Issue-198: Created separate module for network data.
