+++ app/build.gradle
@@ -86,6 +86,8 @@ dependencies {
     testImplementation libraries.mockito
     testImplementation libraries.coroutinesTest
     testImplementation libraries.archTesting
+
+    implementation project(path: ':networkdata')
 }
 
 def getLocalProperty(String propName) {
