+++ app/src/main/java/com/kirchhoff/movies/mapper/core/CoreMapper.kt
@@ -1,12 +1,12 @@
 package com.kirchhoff.movies.mapper.core
 
-import com.kirchhoff.movies.data.network.core.NetworkEntertainmentCredits
-import com.kirchhoff.movies.data.network.core.NetworkEntertainmentPerson
-import com.kirchhoff.movies.data.network.core.NetworkObjectWithName
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentPerson
 import com.kirchhoff.movies.data.ui.core.UIGenre
 import com.kirchhoff.movies.mapper.BaseMapper
+import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentCredits
+import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentPerson
+import com.kirchhoff.movies.networkdata.core.NetworkObjectWithName
 
 class CoreMapper : BaseMapper(), ICoreMapper {
