+++ app/src/main/java/com/kirchhoff/movies/mapper/core/ICoreMapper.kt
@@ -1,9 +1,9 @@
 package com.kirchhoff.movies.mapper.core
 
-import com.kirchhoff.movies.data.network.core.NetworkEntertainmentCredits
-import com.kirchhoff.movies.data.network.core.NetworkObjectWithName
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.core.UIGenre
+import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentCredits
+import com.kirchhoff.movies.networkdata.core.NetworkObjectWithName
 
 interface ICoreMapper {
     fun createUIEntertainmentCredits(credits: NetworkEntertainmentCredits): UIEntertainmentCredits
