+++ app/src/main/java/com/kirchhoff/movies/mapper/discover/DiscoverMapper.kt
@@ -1,12 +1,12 @@
 package com.kirchhoff.movies.mapper.discover
 
-import com.kirchhoff.movies.data.network.core.NetworkPaginated
-import com.kirchhoff.movies.data.network.main.NetworkMovie
-import com.kirchhoff.movies.data.network.main.NetworkTv
 import com.kirchhoff.movies.data.ui.core.UIPaginated
 import com.kirchhoff.movies.data.ui.main.UIMovie
 import com.kirchhoff.movies.data.ui.main.UITv
 import com.kirchhoff.movies.mapper.BaseMapper
+import com.kirchhoff.movies.networkdata.core.NetworkPaginated
+import com.kirchhoff.movies.networkdata.main.NetworkMovie
+import com.kirchhoff.movies.networkdata.main.NetworkTv
 import com.kirchhoff.movies.repository.Result
 
 class DiscoverMapper : BaseMapper(), IDiscoverMapper {
