+++ app/src/main/java/com/kirchhoff/movies/mapper/movie/IMovieDetailsMapper.kt
@@ -1,11 +1,11 @@
 package com.kirchhoff.movies.mapper.movie
 
-import com.kirchhoff.movies.data.network.core.NetworkEntertainmentCredits
-import com.kirchhoff.movies.data.network.details.movie.NetworkMovieDetails
-import com.kirchhoff.movies.data.network.details.movie.NetworkTrailersList
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.details.movie.UIMovieDetails
 import com.kirchhoff.movies.data.ui.details.movie.UITrailersList
+import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentCredits
+import com.kirchhoff.movies.networkdata.details.movie.NetworkMovieDetails
+import com.kirchhoff.movies.networkdata.details.movie.NetworkTrailersList
 import com.kirchhoff.movies.repository.Result
 
 interface IMovieDetailsMapper {
