+++ app/src/main/java/com/kirchhoff/movies/mapper/person/details/IPersonDetailsMapper.kt
@@ -1,9 +1,9 @@
 package com.kirchhoff.movies.mapper.person.details
 
-import com.kirchhoff.movies.data.network.details.person.NetworkPersonCredits
-import com.kirchhoff.movies.data.network.details.person.NetworkPersonDetails
 import com.kirchhoff.movies.data.ui.details.person.UIPersonCredits
 import com.kirchhoff.movies.data.ui.details.person.UIPersonDetails
+import com.kirchhoff.movies.networkdata.details.person.NetworkPersonCredits
+import com.kirchhoff.movies.networkdata.details.person.NetworkPersonDetails
 import com.kirchhoff.movies.repository.Result
 
 interface IPersonDetailsMapper {
