+++ app/src/main/java/com/kirchhoff/movies/mapper/person/details/PersonDetailsMapper.kt
@@ -1,14 +1,14 @@
 package com.kirchhoff.movies.mapper.person.details
 
-import com.kirchhoff.movies.data.network.details.person.NetworkPersonCastCredit
-import com.kirchhoff.movies.data.network.details.person.NetworkPersonCredits
-import com.kirchhoff.movies.data.network.details.person.NetworkPersonCrewCredit
-import com.kirchhoff.movies.data.network.details.person.NetworkPersonDetails
 import com.kirchhoff.movies.data.ui.details.person.UIMediaType
 import com.kirchhoff.movies.data.ui.details.person.UIPersonCredit
 import com.kirchhoff.movies.data.ui.details.person.UIPersonCredits
 import com.kirchhoff.movies.data.ui.details.person.UIPersonDetails
 import com.kirchhoff.movies.mapper.BaseMapper
+import com.kirchhoff.movies.networkdata.details.person.NetworkPersonCastCredit
+import com.kirchhoff.movies.networkdata.details.person.NetworkPersonCredits
+import com.kirchhoff.movies.networkdata.details.person.NetworkPersonCrewCredit
+import com.kirchhoff.movies.networkdata.details.person.NetworkPersonDetails
 import com.kirchhoff.movies.repository.Result
 
 class PersonDetailsMapper : BaseMapper(),
