+++ app/src/main/java/com/kirchhoff/movies/mapper/person/main/PersonsMapper.kt
@@ -1,10 +1,10 @@
 package com.kirchhoff.movies.mapper.person.main
 
-import com.kirchhoff.movies.data.network.core.NetworkPaginated
-import com.kirchhoff.movies.data.network.main.NetworkPerson
 import com.kirchhoff.movies.data.ui.core.UIPaginated
 import com.kirchhoff.movies.data.ui.main.UIPerson
 import com.kirchhoff.movies.mapper.BaseMapper
+import com.kirchhoff.movies.networkdata.core.NetworkPaginated
+import com.kirchhoff.movies.networkdata.main.NetworkPerson
 import com.kirchhoff.movies.repository.Result
 
 class PersonsMapper : BaseMapper(), IPersonsMapper {
