+++ app/src/main/java/com/kirchhoff/movies/mapper/tv/ITvDetailsMapper.kt
@@ -1,9 +1,9 @@
 package com.kirchhoff.movies.mapper.tv
 
-import com.kirchhoff.movies.data.network.core.NetworkEntertainmentCredits
-import com.kirchhoff.movies.data.network.details.tv.NetworkTvDetails
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.details.tv.UITvDetails
+import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentCredits
+import com.kirchhoff.movies.networkdata.details.tv.NetworkTvDetails
 import com.kirchhoff.movies.repository.Result
 
 interface ITvDetailsMapper {
