+++ app/src/main/java/com/kirchhoff/movies/mapper/tv/TvDetailsMapper.kt
@@ -1,11 +1,11 @@
 package com.kirchhoff.movies.mapper.tv
 
-import com.kirchhoff.movies.data.network.core.NetworkEntertainmentCredits
-import com.kirchhoff.movies.data.network.details.tv.NetworkTvDetails
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.details.tv.UITvDetails
 import com.kirchhoff.movies.mapper.BaseMapper
 import com.kirchhoff.movies.mapper.core.ICoreMapper
+import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentCredits
+import com.kirchhoff.movies.networkdata.details.tv.NetworkTvDetails
 import com.kirchhoff.movies.repository.Result
 
 class TvDetailsMapper(private val coreMapper: ICoreMapper) : BaseMapper(), ITvDetailsMapper {
