+++ app/src/main/java/com/kirchhoff/movies/network/services/MovieService.kt
@@ -1,11 +1,11 @@
 package com.kirchhoff.movies.network.services
 
-import com.kirchhoff.movies.data.network.core.NetworkEntertainmentCredits
-import com.kirchhoff.movies.data.network.core.NetworkPaginated
-import com.kirchhoff.movies.data.network.details.movie.NetworkMovieDetails
-import com.kirchhoff.movies.data.network.details.movie.NetworkTrailersList
-import com.kirchhoff.movies.data.network.details.review.NetworkReview
-import com.kirchhoff.movies.data.network.main.NetworkMovie
+import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentCredits
+import com.kirchhoff.movies.networkdata.core.NetworkPaginated
+import com.kirchhoff.movies.networkdata.details.movie.NetworkMovieDetails
+import com.kirchhoff.movies.networkdata.details.movie.NetworkTrailersList
+import com.kirchhoff.movies.networkdata.details.review.NetworkReview
+import com.kirchhoff.movies.networkdata.main.NetworkMovie
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Path
