+++ app/src/test/java/com/kirchhoff/movies/mapper/core/CoreMapperTest.kt
@@ -1,11 +1,11 @@
 package com.kirchhoff.movies.mapper.core
 
-import com.kirchhoff.movies.data.network.core.NetworkEntertainmentCredits
-import com.kirchhoff.movies.data.network.core.NetworkEntertainmentPerson
-import com.kirchhoff.movies.data.network.core.NetworkObjectWithName
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentPerson
 import com.kirchhoff.movies.data.ui.core.UIGenre
+import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentCredits
+import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentPerson
+import com.kirchhoff.movies.networkdata.core.NetworkObjectWithName
 import com.kirchhoff.movies.utils.nextString
 import kotlin.random.Random
 import org.junit.Assert.assertEquals
