+++ settings.gradle
@@ -1,2 +1,3 @@
 include ':app'
+include ':networkdata'
 rootProject.name='Movies'
